re-architected playbook
I wasn't happy with my setup here so I tried NixOS on Phthalo, but found it to be overwhelming to set up.

Since then I reinstalled Debian on that host and decided to keep all my Docker app data in persistent storage so I don't need to jump through hoops to re-configure everything. I also implemented a Unifi role to get that out of the way.

I cleaned up the majority of the playbook and I'm pretty happy with it now. It relies less on homemade roles when I could have been using tried+true community ones. Don't reinvent the wheel.